fix order item not found

+++ OrderService/src/main/java/com/fp/OrderService/repository/OrderItemRepository.java
@@ -10,5 +10,5 @@ import reactor.core.publisher.Mono;
 public interface OrderItemRepository extends ReactiveCrudRepository<OrderItem,Long> {
     Flux<OrderItem>findAllByOrderId(Long orderId);
     @Query("SELECT SUM(o.price) FROM order_item o WHERE o.order_id = :orderId")
-    Mono<Double> sumPriceByOrderId(@Param("orderId") Long orderId);
+    Mono<Float> sumPriceByOrderId(@Param("orderId") Long orderId);
 }
